fix(push): count the configs the delivery budget never reaches
At the shipped defaults — `max_push_configs_per_task` 100,
`push_delivery_timeout` 5s, a hardcoded 30s per-event budget — a webhook
estate that is timing out receives 6 of 100 configured notifications per
event. Measured, not reasoned: the new test asserts 6 delivered and 94
skipped.

The other 94 were reported by one `trace_warn!`, and `trace_warn!`
expands to nothing without the non-default `tracing` feature. In the
build most people run, 94 webhooks going unnotified produced exactly
zero observable output. Each skipped config now increments
`push_outcome::SKIPPED`, which is distinct from `TIMEOUT` because the
two call for different responses: a timeout points at one webhook, a run
of skips points at the arithmetic between the three limits above.

Removed with it: a `tokio::sync::Semaphore::new(16)` built per event,
under a comment describing it as a cap on "hundreds of concurrent HTTP
requests" — a mitigation for a shape this loop has never had, as the
comment two lines below it admitted ("deliveries are currently
sequential"). Scaffolding that describes a protection nobody has is
worse than no comment, because it answers the question a reader came to
ask. What replaces it is the arithmetic that does hold:

    configs_reached == min(len, budget / push_delivery_timeout)

Making delivery genuinely concurrent would raise that ceiling and needs
a joiner this crate does not depend on, so it is recorded as work rather
than done badly here.

The test runs on paused time, so a 30-second budget costs 0.05s of wall
clock. Addendum 4's rule is that paused time and a *real socket* do not
mix; there is no socket here, only `sleep`. Proven able to fail:
deleting the counting loop gives "6 delivered, 0 skipped" against an
expected 100.

`push_delivery.rs` crossed the 500-line ratchet when the test landed and
was split rather than exempted.
