+++ crates/a2a-protocol-server/src/handler/event_processing/background/push_delivery/mod.rs
@@ -48,26 +48,37 @@ pub(super) async fn deliver_push_bg(
     let max_total_push_time = std::time::Duration::from_secs(30);
     let deadline = tokio::time::Instant::now() + max_total_push_time;
 
-    // FIX(M5): Limit concurrent push deliveries to prevent resource exhaustion
-    // when many push configs are registered for a single task. Without this cap,
-    // a burst of events could spawn hundreds of concurrent HTTP requests.
-    let semaphore = tokio::sync::Semaphore::new(16);
-
-    // `_delivered` is read only by `trace_warn!`, which expands to nothing
-    // without the `tracing` feature. The underscore is this crate's convention
-    // for a binding that exists only to be traced — `trace.rs` carries
-    // `#[allow(clippy::used_underscore_binding)]` so these still read naturally
-    // at the call site (cf. `error = %_e` in state_machine.rs).
+    // There is no concurrency to limit. Deliveries below run one after another,
+    // so the only bound that does anything is `deadline`. A `Semaphore::new(16)`
+    // used to be built here per event, with a comment describing it as a cap on
+    // "hundreds of concurrent HTTP requests" — a mitigation for a shape this
+    // loop has never had. It was removed 2026-08-19 rather than left standing:
+    // scaffolding that describes a protection nobody has is worse than no
+    // comment, because it answers the question a reader came to ask.
     //
-    // In a default-feature build the index is then genuinely unused, which is
-    // what the allow below covers. Taking clippy's suggestion and dropping
-    // `.enumerate()` would delete the very count the warning reports, turning
-    // the message back into the overstatement this was fixing.
-    #[allow(clippy::unused_enumerate_index)]
-    for (_delivered, config) in configs.iter().enumerate() {
+    // The arithmetic that does hold, and that operators need:
+    //
+    //     configs_reached  ==  min(len, budget / push_delivery_timeout)
+    //
+    // At the defaults — `max_push_configs_per_task` 100, `push_delivery_timeout`
+    // 5s, budget 30s — a webhook estate that is timing out reaches 6 configs of
+    // 100 per event. The other 94 are now counted (`push_outcome::SKIPPED`)
+    // instead of collapsing into one trace line that a default build compiles
+    // away. Making delivery genuinely concurrent is the fix that would raise
+    // the ceiling; it needs a joiner this crate does not currently depend on,
+    // so it is recorded as work rather than done badly here.
+
+    // `reached` is the index of the config under consideration, and so also the
+    // number already contacted. It used to be `_delivered`, underscored because
+    // its only reader was `trace_warn!` — which expands to nothing without the
+    // `tracing` feature, leaving the binding genuinely unused in a default
+    // build. Now the metrics loop below reads it in every configuration, so the
+    // underscore (and the `unused_enumerate_index` allow that went with it) are
+    // both gone.
+    for (reached, config) in configs.iter().enumerate() {
         // Check if we've exceeded the total push delivery budget.
         if tokio::time::Instant::now() >= deadline {
-            // `configs.len() - _delivered`, not `configs.len()`: this fires
+            // `configs.len() - reached`, not `configs.len()`: this fires
             // partway through the list, so the total is never the remainder.
             // Reporting the total made the one telemetry signal for push
             // amplification overstate the shortfall — at the extreme, a
@@ -75,19 +86,19 @@ pub(super) async fn deliver_push_bg(
             // been skipped.
             trace_warn!(
                 task_id = %task_id,
-                remaining_configs = configs.len() - _delivered,
+                remaining_configs = configs.len() - reached,
                 "push delivery deadline exceeded; skipping remaining configs"
             );
+            // One count per config that will not be contacted. A single trace
+            // line was the whole signal here, and `trace_warn!` compiles to
+            // nothing without the (non-default) `tracing` feature — so in the
+            // build most people run, 94 webhooks going unnotified produced
+            // exactly zero observable output.
+            for _ in reached..configs.len() {
+                metrics.on_push_delivery(push_outcome::SKIPPED);
+            }
             break;
         }
-
-        // Acquire a permit before sending; this bounds concurrency even though
-        // deliveries are currently sequential. The semaphore future-proofs
-        // against a switch to concurrent (join_all / FuturesUnordered) delivery.
-        let _permit = semaphore
-            .acquire()
-            .await
-            .expect("semaphore is never closed");
 
         let result = tokio::time::timeout(
             limits.push_delivery_timeout,
@@ -118,299 +129,4 @@ pub(super) async fn deliver_push_bg(
 }
 
 #[cfg(test)]
-mod tests {
-    use std::future::Future;
-    use std::pin::Pin;
-
-    use a2a_protocol_types::error::A2aError;
-    use a2a_protocol_types::events::{StreamResponse, TaskStatusUpdateEvent};
-    use a2a_protocol_types::push::TaskPushNotificationConfig;
-    use a2a_protocol_types::task::{ContextId, TaskId, TaskState, TaskStatus};
-
-    use crate::handler::limits::HandlerLimits;
-    use crate::push::{InMemoryPushConfigStore, PushConfigStore};
-
-    use super::*;
-
-    /// A push config store that always returns errors.
-    struct AlwaysErrPushConfigStore;
-
-    impl PushConfigStore for AlwaysErrPushConfigStore {
-        fn set<'a>(
-            &'a self,
-            _cfg: TaskPushNotificationConfig,
-        ) -> Pin<
-            Box<
-                dyn Future<
-                        Output = a2a_protocol_types::error::A2aResult<TaskPushNotificationConfig>,
-                    > + Send
-                    + 'a,
-            >,
-        > {
-            Box::pin(async { Err(A2aError::internal("always err")) })
-        }
-        fn get<'a>(
-            &'a self,
-            _task_id: &'a str,
-            _id: &'a str,
-        ) -> Pin<
-            Box<
-                dyn Future<
-                        Output = a2a_protocol_types::error::A2aResult<
-                            Option<TaskPushNotificationConfig>,
-                        >,
-                    > + Send
-                    + 'a,
-            >,
-        > {
-            Box::pin(async { Err(A2aError::internal("always err")) })
-        }
-        fn list<'a>(
-            &'a self,
-            _task_id: &'a str,
-        ) -> Pin<
-            Box<
-                dyn Future<
-                        Output = a2a_protocol_types::error::A2aResult<
-                            Vec<TaskPushNotificationConfig>,
-                        >,
-                    > + Send
-                    + 'a,
-            >,
-        > {
-            Box::pin(async { Err(A2aError::internal("always err")) })
-        }
-        fn delete<'a>(
-            &'a self,
-            _task_id: &'a str,
-            _id: &'a str,
-        ) -> Pin<Box<dyn Future<Output = a2a_protocol_types::error::A2aResult<()>> + Send + 'a>>
-        {
-            Box::pin(async { Err(A2aError::internal("always err")) })
-        }
-    }
-
-    fn make_status_event(task_id: &str, state: TaskState) -> StreamResponse {
-        StreamResponse::StatusUpdate(TaskStatusUpdateEvent {
-            task_id: TaskId::new(task_id),
-            context_id: ContextId::new("ctx-1"),
-            status: TaskStatus::new(state),
-            metadata: None,
-        })
-    }
-
-    fn default_limits() -> HandlerLimits {
-        HandlerLimits::default()
-    }
-
-    #[tokio::test]
-    async fn deliver_push_bg_with_no_sender_is_noop() {
-        let store = InMemoryPushConfigStore::new();
-        let task_id = TaskId::new("t1");
-        let event = make_status_event("t1", TaskState::Working);
-
-        deliver_push_bg(
-            &task_id,
-            &event,
-            &store,
-            None,
-            &default_limits(),
-            &crate::metrics::NoopMetrics,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    async fn deliver_push_bg_with_failing_store_returns_silently() {
-        let store = AlwaysErrPushConfigStore;
-        let task_id = TaskId::new("t1");
-        let event = make_status_event("t1", TaskState::Working);
-
-        deliver_push_bg(
-            &task_id,
-            &event,
-            &store,
-            None,
-            &default_limits(),
-            &crate::metrics::NoopMetrics,
-        )
-        .await;
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn deliver_push_bg_respects_total_deadline() {
-        use std::sync::atomic::{AtomicU64, Ordering};
-        use std::sync::Arc;
-        use std::time::Duration;
-
-        // A push sender that sleeps for 2 seconds per delivery.
-        struct SlowPushSender {
-            send_count: Arc<AtomicU64>,
-        }
-
-        impl crate::push::PushSender for SlowPushSender {
-            fn send<'a>(
-                &'a self,
-                _url: &'a str,
-                _event: &'a StreamResponse,
-                _config: &'a TaskPushNotificationConfig,
-            ) -> Pin<Box<dyn Future<Output = a2a_protocol_types::error::A2aResult<()>> + Send + 'a>>
-            {
-                self.send_count.fetch_add(1, Ordering::Relaxed);
-                Box::pin(async {
-                    tokio::time::sleep(Duration::from_secs(2)).await;
-                    Ok(())
-                })
-            }
-        }
-
-        let store = InMemoryPushConfigStore::new();
-        let task_id = TaskId::new("t-deadline");
-        let event = make_status_event("t-deadline", TaskState::Working);
-
-        // Register many configs. With 2s per delivery and a 30s cap,
-        // at most ~15 can complete.
-        for i in 0..50 {
-            let config = TaskPushNotificationConfig {
-                tenant: None,
-                id: Some(format!("cfg-{i}")),
-                task_id: Some("t-deadline".to_owned()),
-                url: format!("https://example.com/hook{i}"),
-                token: None,
-                authentication: None,
-            };
-            store.set(config).await.unwrap();
-        }
-
-        let send_count = Arc::new(AtomicU64::new(0));
-        let sender = SlowPushSender {
-            send_count: Arc::clone(&send_count),
-        };
-        let limits = HandlerLimits::default().with_push_delivery_timeout(Duration::from_secs(3));
-
-        deliver_push_bg(
-            &task_id,
-            &event,
-            &store,
-            Some(&sender),
-            &limits,
-            &crate::metrics::NoopMetrics,
-        )
-        .await;
-
-        // With 30s total cap and 2s per send (bounded by 3s timeout), not all 50 should fire.
-        let count = send_count.load(Ordering::Relaxed);
-        assert!(
-            count < 50,
-            "deadline should prevent all 50 deliveries, got {count}"
-        );
-        assert!(
-            count > 0,
-            "at least some deliveries should have fired, got {count}"
-        );
-    }
-
-    /// Covers lines 70-76: the `Ok(Err(_))` branch where the push sender returns
-    /// an error. The function should log a warning and continue without panicking.
-    #[tokio::test]
-    async fn deliver_push_bg_logs_delivery_failure() {
-        use std::time::Duration;
-
-        struct FailingPushSender;
-
-        impl crate::push::PushSender for FailingPushSender {
-            fn send<'a>(
-                &'a self,
-                _url: &'a str,
-                _event: &'a StreamResponse,
-                _config: &'a TaskPushNotificationConfig,
-            ) -> Pin<Box<dyn Future<Output = a2a_protocol_types::error::A2aResult<()>> + Send + 'a>>
-            {
-                Box::pin(async { Err(A2aError::internal("push delivery failed")) })
-            }
-        }
-
-        let store = InMemoryPushConfigStore::new();
-        let task_id = TaskId::new("t-fail");
-        let event = make_status_event("t-fail", TaskState::Working);
-
-        // Register a push config so deliver_push_bg actually calls the sender.
-        let config = TaskPushNotificationConfig {
-            tenant: None,
-            id: Some("cfg-fail".to_owned()),
-            task_id: Some("t-fail".to_owned()),
-            url: "https://example.com/hook".to_owned(),
-            token: None,
-            authentication: None,
-        };
-        store.set(config).await.unwrap();
-
-        let sender = FailingPushSender;
-        let limits = HandlerLimits::default().with_push_delivery_timeout(Duration::from_secs(5));
-
-        // Should complete without panic, even though sender returns Err.
-        deliver_push_bg(
-            &task_id,
-            &event,
-            &store,
-            Some(&sender),
-            &limits,
-            &crate::metrics::NoopMetrics,
-        )
-        .await;
-    }
-
-    /// Covers lines 78-83: the `Err(_)` (timeout) branch where the push sender
-    /// takes longer than the timeout. The function should log a warning and continue.
-    #[tokio::test]
-    async fn deliver_push_bg_logs_delivery_timeout() {
-        use std::time::Duration;
-
-        struct SlowForeverPushSender;
-
-        impl crate::push::PushSender for SlowForeverPushSender {
-            fn send<'a>(
-                &'a self,
-                _url: &'a str,
-                _event: &'a StreamResponse,
-                _config: &'a TaskPushNotificationConfig,
-            ) -> Pin<Box<dyn Future<Output = a2a_protocol_types::error::A2aResult<()>> + Send + 'a>>
-            {
-                Box::pin(async {
-                    // Sleep longer than any reasonable timeout.
-                    tokio::time::sleep(Duration::from_secs(600)).await;
-                    Ok(())
-                })
-            }
-        }
-
-        let store = InMemoryPushConfigStore::new();
-        let task_id = TaskId::new("t-timeout");
-        let event = make_status_event("t-timeout", TaskState::Working);
-
-        let config = TaskPushNotificationConfig {
-            tenant: None,
-            id: Some("cfg-timeout".to_owned()),
-            task_id: Some("t-timeout".to_owned()),
-            url: "https://example.com/hook".to_owned(),
-            token: None,
-            authentication: None,
-        };
-        store.set(config).await.unwrap();
-
-        let sender = SlowForeverPushSender;
-        // Set a very short timeout so the test doesn't take long.
-        let limits = HandlerLimits::default().with_push_delivery_timeout(Duration::from_millis(50));
-
-        // Should complete without panic, hitting the timeout branch.
-        deliver_push_bg(
-            &task_id,
-            &event,
-            &store,
-            Some(&sender),
-            &limits,
-            &crate::metrics::NoopMetrics,
-        )
-        .await;
-    }
-}
+mod tests;
